chore: added biobank mutation tests
